+++ packages/ntask/Sources/NTaskLib/NTaskCommand.swift
@@ -1,7 +1,7 @@
 import ArgumentParser
 
 public enum NTaskVersion {
-    public static let current = "0.2.1"
+    public static let current = "0.3.0"
 }
 
 public struct NTaskCommand: AsyncParsableCommand {